eth-decode-transactions/requirements.txt
web3
requests
pandas
numpy
python-dotenv
google-cloud-bigquery